requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
dataclasses-json==0.6.4
deepgram-sdk==3.2.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
python-dotenv==1.0.1
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.10.0
verboselogs==1.7
websockets==12.0
yarl==1.9.4
